Simplify Parser Implementation
Completes the reorganisation of the parser and expression tree. Parsing
is now self-contained in a single file.

Parsing has been moved away from using the parselet structs and traits
along with their related parselet data types and directly into method
implementations for `Token`. These methods are private to the `parse`
module so shouldn't interferere with anything.

The parser itself has been reworked to allow tokens to easily call back
in to advance the iterator and perform look-ahead checks.

Still TODO:

  * Update the error struct to be more descriptive.
  * Update the expression struct to contain metadata about the span of
   	the expression.
  * Update tokens to split them into two sets: basic lexical tokens and
  	trivia tokens.