Ok, I Think the SEGFAULTs are Under Control
Move to use `InContext` variants of functions where possible to enable
two `JitEvaluators` to be run at the same time. Update the
initialisation to be atomic so we don't re-initialise part way through a
test run.

This seems to fix the random SEGFAULTs and or hangs which happened when
more than one test with the JitEvaulator was run at the same time.