Add File Execution Mode
Adds support for running files. In this mode we read the whole file
into a string and then eval each of the expressions out of it rather
than reading it line-by-line. It's still a bit hacky. A nicer solution
would be to have a trait which is shared by both strings and files
that we implement the tokeniser on top of. This works for now though.

The main issue with this is that as each expression is evaluated in a
separate module variables are still not really much use (just a source
of segfaults).