Make Compilation Produce a Module
Stubs out the compilation `emit` function to create an LLVM
Module. The module is wrapped in a Rust object to ensure that it is
disposed of when we're done with it. This commit also moves a
lightly-reimagined version of `low_loader` to the top level of the
crate. In the future this more ergonomic wrapper around LLVM might be
best split off into a separate crate of its own.