Expression Parsing Tweaks
Makes the entry points for expression and expression tree parsing a
bit easier to use. This means that the parser tests can use a
public parser entry point for testing which is nice. Still needs work
on the compilation side to accept single expressions representing
trees rather than `Vec`s.