Fixup Lowering of Expressions
After the move to 'everything is an expression' lowering of function
decls. was broken. Update the `lower_expression` code to add a new
`add_decls` path which pre-adds the declarations for a given
expression recursively. Need to work out how namespacing of functions
will work with this. A problem for Future Will though! Test suite
passes again now which is nice.