Moving to Lowering Semantic Expressions
Updating the interface to the lower module to support lowering
semantic expressions rather than syntax ones. New expressions can be
ported from the syntax tree to the semantic one by adding them to the
transform funcion. Need to work on a grouping class or someting for
the lower parameters though. Passing all that lot around is such a
faff.