Add Semantic Transformation for Identifiers
This starts requiring that we keep more state around when transforming
things. There's no way of knowing the type of an identifier without
resolving it somehow. Need to keep a mapping around and look the
identifier up to find out more metadata about it. This will require a
bit of thinking about as it will define how scoping works.