Initial Work on Mapping Syntax Types
Adds a crude mapping of known syntax types to semantic ones. This is
enough to start clearing up some of the FIXMEs in the translation WRT
function parameter types and local value types. Still needs more work
to correctly tralsate all of the types which the syntax tree is
currently capable of representing.