Add String Literal Support
Now with printing of string literals! This should enable a few more
programs to be written. The string format is in no way fixed at the
moment. Strings should probably be length-prefixed, or length-pointer
pairs. Also would be nice to have a string constants table to preevnt
duplications in the binary.