Add FizzBuzz Spec
Implenets FizzBuzz. This triggered a few bugs in the compiler that
needed to be fixed. Mainly to do with assuming the type of expressions
is `Number`.

Need to work out what to do about `if` expressions. Unwieldly ternary
expressions aren't a long term solution and types mis-matching on the
body blocks will be a pain too.