Deal With Errors from All Expressions in Sequence
If a lowering error was encountered in the lowering of a sequence
expression, but wasn't the last expression in the sequence then the
error would be silently ignored. Fix that by manually keeping track of
what the last expression was when lowering so we can call each one
with the `?` operator.

Part of debugging #10