Add Calling Conventions to Functions
Update the lowering to lower *Ullage* functions as fastcall and the C
interop declarations as C Call. This should allow for faster argument
passing to Ullage functions.

Still TODO with this: allow functions to specify their own calling
convention. This will probably be in the form of a literal string
between the `fun` keyword and the identifier. This can then be used to
define C interop in Ullage itself.