Start Storign String Values in Structures
This doesn't handle printing strings out yet, but values are stored in
`struct` instances and there doesn't seem to be any segfaulting right
now.

The current layout of a `String` is `<{i64, i8*}>`. Ideally we would
move to using pointers to a shared string buffer instead of these. It
looks like that would make printing strings easier too. Need to work
on LLVM array types to allow that however. The proposed format going
forward is `string = <{i32, [0 x i8]}>*`.