Initial `SyntaxText` Implementation
This maps source positions onto `(line, colum)` pairs. The idea is
that the tokeniser will consume characters from this buffer and pass
arround byte offsets. When the parser or compiler encounters and error
the offset can be converted into a useful position to display in an
error message.

For the example soruce text the offsets seem to match up with the
cursor position for Emacs at least. Not sure how well it will handle
Windows line endings. I guess we will ignore the `\r` for line
position. Maybe that's good enough.