Move Syntax Tree Types to Module
Separates out the syntax tree from the rest of the syntax module. This
means that the expression tree is now no longer just chilling in the
`syntax.rs` file and intead gets a shiny file all to itself. The
structure of this tree could well change as the new tokeniser is
developed and it makes sense to have things segragated.