Use an Iterator Adapter to Filter Trivia Tokens
This switches the logic for removing whitespace from being in the
tokeniser itself into a separate iterator adapter. This should
separate the concerns of token production from token
consumption. Hopefully in the future we won't need to filter out the
trivia tokens as the parser will still need them in the tree.