Use `Ident` in AST
Update the `syntax::tree` to refer to things with the interned `Ident`
rather than `String`s. This should make variable lookups simpler. The
semantic tree is still left using `String`s. Only passing the `Ident`
into the tree for time being. This means we loose the rest of the word
token in places.