Be More Accpeting of Optimisation Flags
Use a custom deserialiser to convert the optimisation flags from
strings or numbers into a half-way representation of either a request
for a numeric optimistaion level or a size optimistaion. This should
allow the compiler to accept flags like `-O4` and treat it just the
same as `-O3`. In addition we handle both `-Os and `-Osize`.