Stub Missing Identifiers
If an identifier is expected when parsing but can't be found we now
create a stub one and move on. This still leads to a parse error as
the diagnostic is raised at the point the identifier mis-match wash
found. More of the source can then be parsed and further errors
raised.