Make Spec Runner Deal with Infinite Compilation
As work continues on infallable parsing this becomes more likely. The
change adds a timer which is fired if the compilation process takes
longer than 5 seconds. The timer kills the compiler and causes the
test to fail. As a side effect we will also treat the process being
killed by any signal as a test failure.