Initial `SyntaxTree` Implementation
Start work on the `SyntaxTree`. This will be the main entry point to
the `syntax` module. The plan is to expose static `parse` methods
which take care of constructing the source text, creating a parser,
and collecting the parsed expressions into a tree.

The tree will hold both the parsed source and a set of diagnostics
emitted during the parse. Currently this is just a `Vec<String>` but
hopefully that will change later.