Report Diagnostics from Functions in Parent Scope
Errors from binding a function are pushed into the parent scope once
the bind is complete. This stops us from trying to lower a tree with
`Expression::Error` items inside function definitions.

Now that function and variable declarations are both looked up by the
binder we need to unify the error message for them.