Function Call Return Type Tracking
Call expressions check the return type of the function. Function tpes
contain an ID pointing to the called item which allows the return type
to be looked up. It's all kind of roundabout as we don't have a proper
type information structure to hold centralised type info yet. Ideally
we would only store IDs of types on the expressoin tree and rich type
data in some main state structure.