Add Tests for Mistyped Variable Initialisers
The error message for assigning the wrong value to a variable in an
initialiser wasn't tested. Adds coverage for that case.

This also fixes an issue with the test suite where expectations with
the `!>` appearing in the hint text were always passing. Fix that by
ignoring lines which match the expectation pattern when searching the
error output.