Bail from `Parser::delimited` on No Progress
Add some tests for malformed calls with invalid `,` counts. This
exposes an issue where call expressions with the wrong delimiters at
the end would result in an infinite loop of stubbing out expressions
witout making process. This fixes the loop by keeping track of the
position in the file we are at and bailing from the loop if no
progress is made.