Backpedal
Revert back to LLVM 9. It is still a version update, just not as big a
jump as hoped. :-/

The `clang` toolchain on macOS, along with the build servers, doesn't
seem capable of linking the artifacts produced by LLVM 10 for some
reason. It appears that parmeter names are now a thing meaning the IL
is not compatible. Looks like bitcode attribute changes mean `.bc`
linking isn't an option either.