Use Linker Exectuable in Link
Prepping for being able to choose a different linker exectuable. In
the future when custom linkers are supported and platform default
linking is supported the linker itself will probably want to build the
whole command rather than just providing the exectuable stirng to be
used. This works for now though.